requirements.txt
pandas==2.1.4
numpy==1.24.3
yfinance==0.2.33
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
python-dotenv==1.0.0
openai==1.3.7
requests==2.31.0
plotly==5.18.0
dash==2.18.2
beautifulsoup4==4.12.3
requests-html==0.10.0
feedparser==6.0.11
